Introduce mode and compression param resolution
Adds mode and compression based parameter resolution. With this, if a
user specifies the mode and/or compression params, we will create a
default configuration with the aim of meeting those hints.

Currently, it does not contain support for overriding any of the
parameters. This will be taken in a future commit.

Signed-off-by: John Mazanec <jmazane@amazon.com>